requirements.txt
streamlit
requests
python-slugify
PyYAML
Jinja2
markdown
openai
schedule
